requirements-python.base.txt
# Base Python requirements for UNICEF Analytics Toolkit
# Safe core for data science and development

# Core data science stack
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
statsmodels>=0.14.0
scikit-learn>=1.3.0

# Data manipulation & IO
polars>=0.19.0
openpyxl>=3.1.0
xlrd>=2.0.1
xlsxwriter>=3.1.0
pyreadstat>=1.2.0
pyarrow>=12.0.0

# YAML & configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0
configparser>=6.0.0
toml>=0.10.2

# Jupyter & notebooks
jupyter>=1.0.0
jupyterlab>=4.0.0
notebook>=7.0.0
ipython>=8.14.0
ipykernel>=6.25.0
ipywidgets>=8.1.0
nbconvert>=7.7.0
nbformat>=5.9.0

# Web & HTTP
requests>=2.31.0
urllib3>=2.0.0,<3.0.0
httpx>=0.24.0
aiohttp>=3.8.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Reporting & documentation
jinja2>=3.1.2
reportlab>=4.0.4
weasyprint>=59.0
tabulate>=0.9.0
python-pptx>=0.6.21
python-docx>=0.8.11

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-xdist>=3.5.0
coverage>=7.3.0
hypothesis>=6.84.0
responses>=0.24.0

# Dev utilities
click>=8.1.7
rich>=13.5.0
tqdm>=4.66.0
python-dateutil>=2.8.2
pytz>=2023.3
loguru>=0.7.0
pydantic>=2.3.0